Feedback Feature #11914 Add JSON-API Basic support to the project
https://redmine.biodiversity.agr.gc.ca/issues/11914

Fixed a bug that was affecting the pcr batch resource repository:
-When no fields are specified for a search, SelectionHandler was
creating a join on the "reactions" field, which was causing duplicate
results because joining on reactions will return a copy of the pcr batch
row for each related reaction.
-I solved this by filtering the automatically included relationship
fields to just the to-one fields.
-I wrote a test for this fix called
"findAll_whenNoSortSpecified_resultsAreUniqueAndSortedByAscendingId"
which checks that the IDs of the batches returned are 1 to 10 when the
limit is 10 and no sort is specified.