In progress Feature #12840 Authentication with Keycloak
https://redmine.biodiversity.agr.gc.ca/issues/12840

Added basic authentication with Keycloak:
-Added "keycloak-spring-boot-starter" dependency.
-Added a separate WebSecurityConfigurerAdapter for Keycloak that is used
instead when the property "keycloak.enabled" is true.
-Added keycloak.example.yml as an example of a keycloak config.
-Changed a line in ReadableGroupFilterHandler to get the current
account's name that now works with local, ldap and keycloak accounts.
SecurityContextHolder.getContext().getAuthentication().getName() for a
Keycloak accounts returns Keycloak's ID for that account (e.g.
87959e4d-a07a-4083-a383-79149ad9a138).

Visiting a URL for this application redirects you to a Keycloak login
page. Logging in redirects you back to the URL you tried to access.

The application currently does not create a new DB Account record when a
Keycloak user logs in.