Feature 11916 design authentication support (#8)
* Added spring-security-ldap dependency to the pom.
* Added Auth using ldap: Trying to access the API while unauthenticated
gives you the default Spring login page. Logging in gives you full
access to the data.
* Added local authentication (authentication using non-ldap-account record
stored in the database).
* Added SeqdbDaoAuthenticationProvider: Extends
DaoAuthenticationProvider; Authenticates with LDAP when the user is an
LDAP user, and authenticates with a local account record otherwise.
* Added SeqdbUserDetailsService: Loads a Spring UserDetails object from
the SeqDB database given a username String.
* Added "AccountRepository" interface (automatically implemented by
Spring) to find Accounts using less code than JPA criteria queries.
* Made Account-finding by username case-insensitive (Updated the tests to
check for this).
* Added SeqdbLdapUserDetailsMapper + tests, which adds an LDAP user's
local account record + default group when they first log in. (We should
review whether this is needed for the API)
* Added example LDAP yml config.
* Added link to SecurityRepositories that explains Spring Data's
generated repositories.
* Combined the two CountryRepository methods.
* Combined the two ProvinceRepository methods.
* Added uml diagram of the security classes, with png export.
* Added extra note to SeqdbUserDetailsService javadoc about Spring
Security's usage.
* Changed the lombok-generated constructor access level for
SeqdbUserDetailsService and SeqdbLdapUserDetailsMapper to
package-protected.