In progress Feature #12364 Initial authorization based on group(s)
https://redmine.biodiversity.agr.gc.ca/issues/12364

-Rewrote JpaResourceRepository#findOne to call findAll to enable
readable group filtering.
-Added case to ReadableGroupFilterHandler to disable results restriction
for admins.
-Added ReadableGroupFilterHandler tests for searching as admin and for
calling findOne on a resource that exists but belongs to a group not
readable by the user.
-Added ReadableGroupFilterHandlerFactory class, which reduces the amount
of code to create a ReadableGroupFilterHandler.
-Modified WritableGroupAuthorizationAspect#requireGroupAccess to reduce
the amount of 'if' branching.
-Removed WritableGroupAuthorizationAspectIT tets for null accountsgroup
because accountsgroup should never be null; If the user has no read
permission on the group then they will get a 404 error when linking to
a group they have no access to.
-Added error message to findOne's ResourceNotFoundException.