Support #15086 Test ModelMapper and MapStruct using ThermocyclerProfile
https://redmine.biodiversity.agr.gc.ca/issues/15086

> JpaDtoRepository

* Replaced the standard modelmapper with a configured one created from a
base class.

* Added a check to map the hibernate database result to the entity first
if the dto requires mapping. This is done since it is normally returned
as a hashmap and the mapper cannot be configured to map that.

> Mapping Test

*updated the test but this file is only here while researching and will
be deleted if this branch is merged into dev.

> ThermocyclerIT

* created the IT for thermocycler, one test currently fails.
The test that fails is
createPcrProfile_onSucess_allFieldsHavePersistedValues(). The test case
of this is when updating the entities based on DTO, Since only some of
the fields are required to be updated we require mapping on the field
level and cannot use model mapper for that, thus a custom mapper will be
needed.

> SelectionHandler

* removed unused imports